cmd/juju: move utils from status to common
