+++ cmd/juju/status/output_oneline.go
@@ -9,6 +9,7 @@ import (
 	"strings"
 
 	"github.com/juju/errors"
+	"github.com/juju/juju/cmd/juju/common"
 )
 
 // FormatOneline returns a brief list of units and their subordinates.
@@ -60,9 +61,9 @@ func formatOneline(value interface{}, printf onelinePrintf) ([]byte, error) {
 		printf(&out, format, uName, u, level)
 	}
 
-	for _, svcName := range sortStringsNaturally(stringKeysFromMap(fs.Services)) {
+	for _, svcName := range common.SortStringsNaturally(stringKeysFromMap(fs.Services)) {
 		svc := fs.Services[svcName]
-		for _, uName := range sortStringsNaturally(stringKeysFromMap(svc.Units)) {
+		for _, uName := range common.SortStringsNaturally(stringKeysFromMap(svc.Units)) {
 			unit := svc.Units[uName]
 			pprint(uName, unit, 0)
 			recurseUnits(unit, 1, pprint)
